Added generated music sheet
This pdf is the music sheet that will be output from the program.
Information about title, composer, copyright, etc, will be added manually after.

The file.musicxml is the format of the file that can be loaded using e.g. musescore,
and that contains all the information of the music sheet shown.